refactor(cli): decompose readline host into command/turn/context modules
CoderCLI (873 lines) splits along its real seams, behavior preserved:

- readline/command-surface.ts: LOCAL_COMMANDS / HELP_ITEMS /
  RETIRED_COMMANDS / HELP_FOOTER plus routeSlashInput(), the '/'-input
  router (retired notices, built-in > skill > error resolution, /skills
  and /wt transforms). The pre-existing quirk where a transformed
  non-local skill message still falls through to handleCommand (and is
  reported unknown) is preserved verbatim and marked with a NOTE.
- readline/host-commands.ts: ReadlineCommands with the handleCommand
  switch and the /model command (index-first resolution kept).
- readline/agent-turn.ts: executeAgentTurn() — session banner, engine
  run + callbacks, abort-sentinel check, run summary, session save and
  daily-log capture. Queue/abort state stays in the host's input loop.
- readline/host-context.ts: ReadlineHost collaborator interface plus
  shared host helpers (estimateTokens, getKeepLastTurns,
  resolveCurrentSessionId, syncSessionTaskListBinding,
  restoreSessionModel).
- readline/readline-host.ts keeps assembly, the input loop, Esc/SIGINT
  handling and session close. 257/271/154/152/111 lines.

Validated: build, full suite green, tsc error profile identical, smoke
run exercising /help, /model, /status, unknown-command routing and
/exit through the built binary.
